requirements.txt
pandas==1.4.2
Pillow==9.5.0
streamlit==1.24.0
catboost==1.2
numpy==1.23.5
